rename & coin
