+++ app/app.go
@@ -32,9 +32,9 @@ import (
 	"github.com/okex/exchain/libs/cosmos-sdk/x/supply"
 	"github.com/okex/exchain/libs/cosmos-sdk/x/upgrade"
 	"github.com/okex/exchain/libs/iavl"
-	"github.com/okex/exchain/libs/ibc-go/modules/application/transfer"
-	ibctransferkeeper "github.com/okex/exchain/libs/ibc-go/modules/application/transfer/keeper"
-	ibctransfertypes "github.com/okex/exchain/libs/ibc-go/modules/application/transfer/types"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer"
+	ibctransferkeeper "github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/keeper"
+	ibctransfertypes "github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/types"
 	ibc "github.com/okex/exchain/libs/ibc-go/modules/core"
 	ibcclient "github.com/okex/exchain/libs/ibc-go/modules/core/02-client"
 	porttypes "github.com/okex/exchain/libs/ibc-go/modules/core/05-port/types"
