+++ libs/ibc-go/modules/apps/transfer/alias.go
@@ -1,0 +1,11 @@
+package transfer
+
+import (
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/keeper"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/types"
+)
+
+var (
+	NewKeeper = keeper.NewKeeper
+	ModuleCdc = types.ModuleCdc
+)
