+++ libs/ibc-go/modules/apps/transfer/keeper/hooks.go
@@ -2,7 +2,7 @@ package keeper
 
 import (
 	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
-	"github.com/okex/exchain/libs/ibc-go/modules/application/transfer/types"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/types"
 )
 
 // SetHooks sets the hooks for the IBC transfer module
