+++ libs/ibc-go/modules/apps/transfer/keeper/msg_server.go
@@ -3,7 +3,7 @@ package keeper
 import (
 	"context"
 	sdk "github.com/okex/exchain/libs/cosmos-sdk/types"
-	"github.com/okex/exchain/libs/ibc-go/modules/application/transfer/types"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/types"
 )
 
 //var _ types.MsgServer = Keeper{}
@@ -23,10 +23,8 @@ func (k Keeper) Transfer(goCtx context.Context, msg *types.MsgTransfer) (*MsgTra
 	}
 	if err := k.SendTransfer(
 		ctx, msg.SourcePort, msg.SourceChannel, sdk.Coin{
-			Denom: msg.Token.Denom,
-			Amount: sdk.Dec{
-				Int: msg.Token.Amount.BigInt(),
-			},
+			Denom:  msg.Token.Denom,
+			Amount: msg.Token.Amount.ToDec(),
 		}, sender, msg.Receiver, msg.TimeoutHeight, msg.TimeoutTimestamp,
 	); err != nil {
 		return nil, err
