+++ libs/ibc-go/modules/apps/transfer/module.go
@@ -15,10 +15,10 @@ import (
 	"github.com/okex/exchain/libs/cosmos-sdk/types/module"
 	capabilitytypes "github.com/okex/exchain/libs/cosmos-sdk/x/capability/types"
 	simtypes "github.com/okex/exchain/libs/cosmos-sdk/x/simulation"
-	"github.com/okex/exchain/libs/ibc-go/modules/application/transfer/client/cli"
-	"github.com/okex/exchain/libs/ibc-go/modules/application/transfer/keeper"
-	"github.com/okex/exchain/libs/ibc-go/modules/application/transfer/simulation"
-	"github.com/okex/exchain/libs/ibc-go/modules/application/transfer/types"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/client/cli"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/keeper"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/simulation"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/types"
 	channeltypes "github.com/okex/exchain/libs/ibc-go/modules/core/04-channel/types"
 	porttypes "github.com/okex/exchain/libs/ibc-go/modules/core/05-port/types"
 	host "github.com/okex/exchain/libs/ibc-go/modules/core/24-host"
