+++ libs/ibc-go/modules/apps/transfer/simulation/params.go
@@ -1,15 +1,15 @@
 package simulation
 
 import (
-	"github.com/okex/exchain/libs/ibc-go/modules/application/transfer/types"
-	"github.com/okex/exchain/libs/cosmos-sdk/x/simulation"
 	gogotypes "github.com/gogo/protobuf/types"
+	"github.com/okex/exchain/libs/cosmos-sdk/x/simulation"
+	"github.com/okex/exchain/libs/ibc-go/modules/apps/transfer/types"
 	"math/rand"
 )
 
 // ParamChanges defines the parameters that can be modified by param change proposals
 // on the simulation
-func ParamChanges(r *rand.Rand) []simulation.ParamChange{
+func ParamChanges(r *rand.Rand) []simulation.ParamChange {
 	return []simulation.ParamChange{
 		simulation.NewSimParamChange(types.ModuleName, string(types.KeySendEnabled),
 			func(r *rand.Rand) string {
